+++ CMakeLists.txt
@@ -309,10 +309,6 @@ set(SeQuant_src
         SeQuant/core/tag.hpp
         SeQuant/core/tensor_canonicalizer.cpp
         SeQuant/core/tensor_canonicalizer.hpp
-        SeQuant/core/tensor_network_v2.cpp
-        SeQuant/core/tensor_network_v2.hpp
-        SeQuant/core/tensor_network_v3.cpp
-        SeQuant/core/tensor_network_v3.hpp
         SeQuant/core/tensor_network/canonicals.hpp
         SeQuant/core/tensor_network/slot.hpp
         SeQuant/core/tensor_network/typedefs.hpp
@@ -322,6 +318,10 @@ set(SeQuant_src
         SeQuant/core/tensor_network/vertex_painter.hpp
         SeQuant/core/tensor_network/v1.cpp
         SeQuant/core/tensor_network/v1.hpp
+        SeQuant/core/tensor_network/v2.cpp
+        SeQuant/core/tensor_network/v2.hpp
+        SeQuant/core/tensor_network/v3.cpp
+        SeQuant/core/tensor_network/v3.hpp
         SeQuant/core/utility/atomic.hpp
         SeQuant/core/utility/context.hpp
         SeQuant/core/utility/debug.hpp
@@ -370,8 +370,8 @@ set(SeQuant_src
 
 #### disable unity_build for v2:
 set_source_files_properties(
-        SeQuant/core/tensor_network_v2.cpp
-        SeQuant/core/tensor_network_v2.hpp
+        SeQuant/core/tensor_network/v2.cpp
+        SeQuant/core/tensor_network/v2.hpp
         PROPERTIES SKIP_UNITY_BUILD_INCLUSION TRUE
 )
